Student Portal with test cases/requirements.txt
virtualenv==20.4.3
asgiref==3.2.10
Django==3.0.8
django-crispy-forms==1.9.1
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
django-user-unique-email==0.1.1
pytest==6.2.4
